refactor(swift): split oversized files into focused modules
Three files were over 800 lines, mixing multiple concerns:

- ContentView.swift (1170 → 790): keymap handler registration
  moved to ContentView+Keymaps.swift extension. ContentView state
  and helpers are now `internal` (not `private`) so the extension
  can access them.

- EditableWebView.swift (1510 → 356): the 1100-line HTML/JS
  template builder moved to EditableWebViewHTML.swift as a pure
  static enum function. backgroundColor and vimInsertExitKeys
  are now passed explicitly instead of read from self.

- EmailComposition.swift (823 → 214): reply/forward factory
  methods moved to EmailDraftFactory.swift; the local-draft HTTP
  client moved to DraftManager.swift. The model file now contains
  only EmailDraft, EmailAttachment, EmailSendingError, EmailHelper.

No behavioral changes — all tests still pass, build still green.
