+++ macos/durian/ContentView.swift
@@ -26,27 +26,30 @@ enum DetailViewMode: Equatable {
 
 struct ContentView: View {
     @Environment(\.openWindow) private var openWindow
-    @ObservedObject private var accountManager = AccountManager.shared
+    // Note: several properties below are `internal` (no `private`) so that
+    // the keymap handler extension in ContentView+Keymaps.swift can access
+    // them. ContentView itself is still an internal app type.
+    @ObservedObject var accountManager = AccountManager.shared
     @ObservedObject private var keymapsManager = KeymapsManager.shared
-    @ObservedObject private var keymapHandler = KeymapHandler.shared
+    @ObservedObject var keymapHandler = KeymapHandler.shared
     @ObservedObject private var profileManager = ProfileManager.shared
     @ObservedObject private var syncManager = SyncManager.shared
     @ObservedObject private var networkMonitor = NetworkMonitor.shared
     @ObservedObject private var bannerManager = BannerManager.shared
-    @State private var selectedTagID: String? = "inbox"
-    @State private var cursorEmailId: String? = nil       // Highlighted email (cursor position)
-    @State private var markedEmails: Set<String> = []     // Marked emails (selection for batch ops)
-    @State private var detailMode: DetailViewMode = .empty
-    @State private var showSearchPopup: Bool = false
-    @State private var showTagPicker: Bool = false
-    @State private var allTags: [String] = []
-    @State private var visualModeAnchor: String? = nil    // Anchor for visual mode range selection
-    @State private var isSearchMode = false
+    @State var selectedTagID: String? = "inbox"
+    @State var cursorEmailId: String? = nil       // Highlighted email (cursor position)
+    @State var markedEmails: Set<String> = []     // Marked emails (selection for batch ops)
+    @State var detailMode: DetailViewMode = .empty
+    @State var showSearchPopup: Bool = false
+    @State var showTagPicker: Bool = false
+    @State var allTags: [String] = []
+    @State var visualModeAnchor: String? = nil    // Anchor for visual mode range selection
+    @State var isSearchMode = false
     @State private var bodyFetchTask: Task<Void, Never>?
-    @State private var searchResults: [MailMessage] = []
-    @State private var lastSearchQuery = ""
-    @State private var focusedMessageIndex: Int = 0
-    @State private var isThreadFocused: Bool = false
+    @State var searchResults: [MailMessage] = []
+    @State var lastSearchQuery = ""
+    @State var focusedMessageIndex: Int = 0
+    @State var isThreadFocused: Bool = false
 
     var body: some View {
         ZStack {
@@ -493,7 +496,7 @@ struct ContentView: View {
     }
     
     /// Number of thread messages for the currently focused email
-    private var currentThreadMessageCount: Int {
+    var currentThreadMessageCount: Int {
         guard let emailId = cursorEmailId,
               let email = accountManager.mailMessages.first(where: { $0.id == emailId })
                           ?? searchResults.first(where: { $0.id == emailId }),
@@ -592,7 +595,7 @@ struct ContentView: View {
         Task { await accountManager.togglePin(id: emailId) }
     }
 
-    private func toggleRead() {
+    func toggleRead() {
         guard !markedEmails.isEmpty else { return }
         Task {
             await accountManager.toggleReadForMessages(ids: markedEmails)
@@ -609,7 +612,7 @@ struct ContentView: View {
     
     // MARK: - Compose
     
-    private func openNewCompose() {
+    func openNewCompose() {
         guard !ConfigManager.shared.getAccounts().isEmpty else { return }
         let draftId = DraftService.shared.createDraft(from: defaultFromAccount)
         openWindow(value: draftId)
@@ -631,7 +634,7 @@ struct ContentView: View {
         return ConfigManager.shared.getAccounts().first?.email
     }
     
-    private func replyToSelected() {
+    func replyToSelected() {
         guard let email = selectedEmail,
               case .loaded = email.bodyState,
               let fromAccount = defaultFromAccount else { return }
@@ -644,7 +647,7 @@ struct ContentView: View {
         }
     }
 
-    private func replyAllToSelected() {
+    func replyAllToSelected() {
         guard let email = selectedEmail,
               case .loaded = email.bodyState,
               let fromAccount = defaultFromAccount else { return }
@@ -665,7 +668,7 @@ struct ContentView: View {
         return (body: response.body, html: response.html)
     }
     
-    private func forwardSelected() {
+    func forwardSelected() {
         guard let email = selectedEmail,
               case .loaded = email.bodyState,
               let fromAccount = defaultFromAccount else { return }
@@ -701,7 +704,7 @@ struct ContentView: View {
     // MARK: - Navigation Helpers
     
     /// Advance cursor and selection to the given email, or clear if nil.
-    private func advanceCursor(to emailId: String?) {
+    func advanceCursor(to emailId: String?) {
         if let emailId = emailId {
             cursorEmailId = emailId
             markedEmails = [emailId]
@@ -715,7 +718,7 @@ struct ContentView: View {
 
     /// Compute the next email to select after removing the given IDs from the list.
     /// Picks the email just below the lowest removed index, or the one above if at the end.
-    private func nextEmailId(after removedIds: Set<String>) -> String? {
+    func nextEmailId(after removedIds: Set<String>) -> String? {
         let ids = sortedEmailIds
         guard let firstRemovedIndex = ids.firstIndex(where: { removedIds.contains($0) }) else { return nil }
         // Try the email just after the last contiguous removed item
@@ -730,21 +733,21 @@ struct ContentView: View {
     }
 
     /// Get sorted email IDs matching visual order (pinned first, then by timestamp)
-    private var sortedEmailIds: [String] {
+    var sortedEmailIds: [String] {
         let pinned = displayEmails.filter { $0.isPinned }.sorted { $0.timestamp > $1.timestamp }
         let unpinned = displayEmails.filter { !$0.isPinned }.sorted { $0.timestamp > $1.timestamp }
         return (pinned + unpinned).map { $0.id }
     }
     
     /// Get current email index in sorted list (based on cursor position)
-    private func currentEmailIndex() -> Int? {
+    func currentEmailIndex() -> Int? {
         guard let currentId = cursorEmailId else { return nil }
         return sortedEmailIds.firstIndex(of: currentId)
     }
     
     /// Navigate to email at specific index (clamped to valid range)
     /// Updates cursor position and handles visual mode selection
-    private func navigateToEmail(at index: Int) {
+    func navigateToEmail(at index: Int) {
         guard !sortedEmailIds.isEmpty else { return }
         let clampedIndex = max(0, min(index, sortedEmailIds.count - 1))
         let targetId = sortedEmailIds[clampedIndex]
@@ -784,387 +787,4 @@ struct ContentView: View {
         guard clampedStart <= clampedEnd else { return [] }
         return Set(sortedEmailIds[clampedStart...clampedEnd])
     }
-    
-    /// Register all keymap handlers
-    private func registerKeymapHandlers() {
-        // Navigation: j/k with count support (5j, 3k)
-        // In visual mode, extends selection from anchor to cursor
-        keymapHandler.registerHandler(for: .nextEmail) { [self] count in
-            await MainActor.run {
-                if let current = currentEmailIndex() {
-                    navigateToEmail(at: current + count)
-                } else if !sortedEmailIds.isEmpty {
-                    navigateToEmail(at: 0)
-                }
-            }
-        }
-        
-        keymapHandler.registerHandler(for: .prevEmail) { [self] count in
-            await MainActor.run {
-                if let current = currentEmailIndex() {
-                    navigateToEmail(at: current - count)
-                } else if !sortedEmailIds.isEmpty {
-                    navigateToEmail(at: sortedEmailIds.count - 1)
-                }
-            }
-        }
-        
-        // First/Last: gg, G
-        keymapHandler.registerSimpleHandler(for: .firstEmail) { [self] in
-            await MainActor.run {
-                navigateToEmail(at: 0)
-            }
-        }
-        
-        keymapHandler.registerSimpleHandler(for: .lastEmail) { [self] in
-            await MainActor.run {
-                navigateToEmail(at: sortedEmailIds.count - 1)
-            }
-        }
-        
-        // Page navigation: Ctrl+d, Ctrl+u (10 emails per page)
-        let pageSize = 10
-        keymapHandler.registerHandler(for: .pageDown) { [self] count in
-            await MainActor.run {
-                if let current = currentEmailIndex() {
-                    navigateToEmail(at: current + (pageSize * count))
-                } else {
-                    navigateToEmail(at: 0)
-                }
-            }
-        }
-        
-        keymapHandler.registerHandler(for: .pageUp) { [self] count in
-            await MainActor.run {
-                if let current = currentEmailIndex() {
-                    navigateToEmail(at: current - (pageSize * count))
-                } else {
-                    navigateToEmail(at: sortedEmailIds.count - 1)
-                }
-            }
-        }
-        
-        // Search: /
-        keymapHandler.registerSimpleHandler(for: .search) { [self] in
-            await MainActor.run {
-                showSearchPopup = true
-            }
-        }
-        
-        // Tag Picker: t
-        keymapHandler.registerSimpleHandler(for: .tagPicker) { [self] in
-            await MainActor.run {
-                guard !markedEmails.isEmpty else { return }
-                showTagPicker = true
-                Task { allTags = await accountManager.fetchAllTags() }
-            }
-        }
-
-        // Compose: c
-        keymapHandler.registerSimpleHandler(for: .compose) { [self] in
-            await MainActor.run {
-                openNewCompose()
-            }
-        }
-        
-        // Reply: r
-        keymapHandler.registerSimpleHandler(for: .reply) { [self] in
-            await MainActor.run {
-                replyToSelected()
-            }
-        }
-        
-        // Reply All: R (Shift+r)
-        keymapHandler.registerSimpleHandler(for: .replyAll) { [self] in
-            await MainActor.run {
-                replyAllToSelected()
-            }
-        }
-        
-        // Forward: f
-        keymapHandler.registerSimpleHandler(for: .forward) { [self] in
-            await MainActor.run {
-                forwardSelected()
-            }
-        }
-        
-        // View control: q - close detail, search popup, or tag picker (NOT escape - that's for visual mode)
-        keymapHandler.registerSimpleHandler(for: .closeDetail) { [self] in
-            await MainActor.run {
-                if showSearchPopup {
-                    showSearchPopup = false
-                } else if showTagPicker {
-                    showTagPicker = false
-                } else {
-                    detailMode = .empty
-                }
-            }
-        }
-        
-        // Toggle Pin: s
-        keymapHandler.registerSimpleHandler(for: .toggleStar) { [self] in
-            await MainActor.run {
-                guard let emailId = markedEmails.first else { return }
-                Task {
-                    await accountManager.togglePin(id: emailId)
-                }
-            }
-        }
-        
-        // Toggle Read: u - now works with multi-selection
-        keymapHandler.registerSimpleHandler(for: .toggleRead) { [self] in
-            await MainActor.run {
-                toggleRead()
-            }
-        }
-        
-        // Archive: a - add archive tag and remove inbox tag (works with multi-selection)
-        keymapHandler.registerSimpleHandler(for: .archiveEmail) { [self] in
-            let ids = await MainActor.run { () -> Set<String> in
-                let ids = markedEmails
-                guard !ids.isEmpty else { return [] }
-                let next = nextEmailId(after: ids)
-                visualModeAnchor = nil
-                keymapHandler.engine.exitVisualMode()
-                accountManager.removeLocally(ids: ids)
-                advanceCursor(to: next)
-                return ids
-            }
-            for id in ids {
-                await accountManager.modifyTagsWithoutSync(id: id, add: ["archive"], remove: ["inbox"])
-            }
-            await accountManager.syncAndRefresh()
-        }
-
-        // Delete: dd - works with multi-selection
-        keymapHandler.registerSimpleHandler(for: .deleteEmail) { [self] in
-            let ids = await MainActor.run { () -> Set<String> in
-                guard !markedEmails.isEmpty else { return [] }
-                let ids = markedEmails
-                let next = nextEmailId(after: ids)
-                accountManager.removeLocally(ids: ids)
-                visualModeAnchor = nil
-                keymapHandler.engine.exitVisualMode()
-                advanceCursor(to: next)
-                return ids
-            }
-            await accountManager.deleteMessagesWithoutSync(ids: ids)
-            await accountManager.syncAndRefresh()
-        }
-        
-        // ═══════════════════════════════════════════════════════════
-        // FOLDER NAVIGATION HANDLERS
-        // ═══════════════════════════════════════════════════════════
-
-        // gi - Go to Inbox
-        keymapHandler.registerSimpleHandler(for: .goInbox) { [self] in
-            await MainActor.run { selectedTagID = "inbox" }
-        }
-
-        // gs - Go to Sent
-        keymapHandler.registerSimpleHandler(for: .goSent) { [self] in
-            await MainActor.run { selectedTagID = "sent" }
-        }
-
-        // gd - Go to Drafts
-        keymapHandler.registerSimpleHandler(for: .goDrafts) { [self] in
-            await MainActor.run { selectedTagID = "drafts" }
-        }
-
-        // ga - Go to Archive
-        keymapHandler.registerSimpleHandler(for: .goArchive) { [self] in
-            await MainActor.run { selectedTagID = "archive" }
-        }
-
-        // ═══════════════════════════════════════════════════════════
-        // VISUAL MODE HANDLERS
-        // ═══════════════════════════════════════════════════════════
-        
-        // v - Enter LINE visual mode (range selection: anchor to cursor)
-        keymapHandler.registerSimpleHandler(for: .enterVisualMode) { [self] in
-            await MainActor.run {
-                guard keymapHandler.engine.visualModeType == .none else { return }
-                keymapHandler.engine.enterVisualMode(.line)
-                // Set anchor to current cursor position
-                visualModeAnchor = cursorEmailId
-                Log.debug("VISUAL", "Entered LINE mode, anchor: \(visualModeAnchor ?? "nil")")
-            }
-        }
-        
-        // V (Shift+v) - Enter TOGGLE visual mode (individual selection with Space)
-        keymapHandler.registerSimpleHandler(for: .enterToggleMode) { [self] in
-            await MainActor.run {
-                guard keymapHandler.engine.visualModeType == .none else { return }
-                keymapHandler.engine.enterVisualMode(.toggle)
-                // Mark current email initially
-                if let currentId = cursorEmailId {
-                    markedEmails = [currentId]
-                }
-                Log.debug("VISUAL", "Entered TOGGLE mode, initial mark: \(cursorEmailId ?? "nil")")
-            }
-        }
-        
-        // Space - Toggle selection (only in TOGGLE mode)
-        keymapHandler.registerSimpleHandler(for: .toggleSelection) { [self] in
-            await MainActor.run {
-                // Only works in Toggle mode
-                guard keymapHandler.engine.visualModeType == .toggle else { return }
-                guard let currentId = cursorEmailId else { return }
-                
-                // Toggle mark on current cursor position
-                if markedEmails.contains(currentId) {
-                    // Unmark (but keep at least one)
-                    if markedEmails.count > 1 {
-                        markedEmails.remove(currentId)
-                    }
-                } else {
-                    // Mark
-                    markedEmails.insert(currentId)
-                }
-                
-                // Move cursor to next email (mutt-style)
-                if let current = currentEmailIndex(), current < sortedEmailIds.count - 1 {
-                    cursorEmailId = sortedEmailIds[current + 1]
-                    // Selection stays unchanged (toggle mode)
-                }
-            }
-        }
-        
-        // Escape - Exit visual mode (clears selection, keeps cursor)
-        keymapHandler.registerSimpleHandler(for: .exitVisualMode) { [self] in
-            await MainActor.run {
-                // Always clean up visual mode state (engine may have already exited)
-                if visualModeAnchor != nil || keymapHandler.engine.visualModeType != .none {
-                    keymapHandler.engine.exitVisualMode()
-                    if let cursor = cursorEmailId {
-                        markedEmails = [cursor]
-                    }
-                    visualModeAnchor = nil
-                    Log.debug("VISUAL", "Exited visual mode")
-                } else if showTagPicker {
-                    showTagPicker = false
-                } else if showSearchPopup {
-                    showSearchPopup = false
-                } else if isSearchMode {
-                    isSearchMode = false
-                    searchResults = []
-                    lastSearchQuery = ""
-                } else {
-                    detailMode = .empty
-                }
-            }
-        }
-
-        // ═══════════════════════════════════════════════════════════
-        // THREAD CONTEXT HANDLERS
-        // ═══════════════════════════════════════════════════════════
-
-        // l - Enter thread view
-        keymapHandler.registerSimpleHandler(for: .enterThread) { [self] in
-            await MainActor.run {
-                guard cursorEmailId != nil else { return }
-                focusedMessageIndex = 0
-                isThreadFocused = true
-                keymapHandler.engine.setContext(.thread)
-            }
-        }
-
-        // j/k in thread - scroll (supports count: 3j = scroll 3x)
-        keymapHandler.registerHandler(for: .scrollDown, context: .thread) { count in
-            for _ in 0..<count {
-                NotificationCenter.default.post(name: .threadScrollDown, object: nil)
-            }
-        }
-
-        keymapHandler.registerHandler(for: .scrollUp, context: .thread) { count in
-            for _ in 0..<count {
-                NotificationCenter.default.post(name: .threadScrollUp, object: nil)
-            }
-        }
-
-        // Ctrl+d/u in thread - half-page scroll
-        keymapHandler.registerHandler(for: .pageDown, context: .thread) { count in
-            let steps = 5 * count
-            for _ in 0..<steps {
-                NotificationCenter.default.post(name: .threadScrollDown, object: nil)
-            }
-        }
-
-        keymapHandler.registerHandler(for: .pageUp, context: .thread) { count in
-            let steps = 5 * count
-            for _ in 0..<steps {
-                NotificationCenter.default.post(name: .threadScrollUp, object: nil)
-            }
-        }
-
-        // n/N in thread - navigate messages (supports count: 3n = jump 3)
-        keymapHandler.registerHandler(for: .nextMessage, context: .thread) { [self] count in
-            await MainActor.run {
-                let max = currentThreadMessageCount - 1
-                focusedMessageIndex = min(focusedMessageIndex + count, max)
-            }
-        }
-
-        keymapHandler.registerHandler(for: .prevMessage, context: .thread) { [self] count in
-            await MainActor.run {
-                guard focusedMessageIndex > 0 else { return }
-                focusedMessageIndex = max(focusedMessageIndex - count, 0)
-            }
-        }
-
-        // gg/G in thread - first/last message
-        keymapHandler.registerSimpleHandler(for: .firstEmail, context: .thread) { [self] in
-            await MainActor.run {
-                guard focusedMessageIndex > 0 else { return }
-                focusedMessageIndex = 0
-            }
-        }
-
-        keymapHandler.registerSimpleHandler(for: .lastEmail, context: .thread) { [self] in
-            await MainActor.run {
-                let last = max(currentThreadMessageCount - 1, 0)
-                if focusedMessageIndex == last {
-                    NotificationCenter.default.post(name: .threadScrollToBottom, object: nil)
-                }
-                focusedMessageIndex = last
-            }
-        }
-
-        // h/Escape in thread - back to list
-        keymapHandler.registerSimpleHandler(for: .closeDetail, context: .thread) { [self] in
-            await MainActor.run {
-                isThreadFocused = false
-                keymapHandler.engine.setContext(.list)
-            }
-        }
-
-        // r in thread - reply
-        keymapHandler.registerSimpleHandler(for: .reply, context: .thread) { [self] in
-            await MainActor.run {
-                replyToSelected()
-            }
-        }
-
-        // ═══════════════════════════════════════════════════════════
-        // POPUP CONTEXT HANDLERS (search, tag picker)
-        // ═══════════════════════════════════════════════════════════
-
-        for ctx in [KeymapContext.search, KeymapContext.tagPicker] {
-            keymapHandler.registerSimpleHandler(for: .selectNext, context: ctx) {
-                NotificationCenter.default.post(name: .popupSelectNext, object: nil)
-            }
-            keymapHandler.registerSimpleHandler(for: .selectPrev, context: ctx) {
-                NotificationCenter.default.post(name: .popupSelectPrev, object: nil)
-            }
-            keymapHandler.registerSimpleHandler(for: .closePopup, context: ctx) { [self] in
-                await MainActor.run {
-                    showSearchPopup = false
-                    showTagPicker = false
-                }
-            }
-        }
-
-        Log.debug("KEYMAPS", "All handlers registered")
-    }
-    
 }
